requirements.txt
asgiref==3.2.10
autopep8==1.5.4
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.6.0
diff-match-patch==20200713
dj-rest-auth==2.1.3
Django==3.1
django-allauth==0.44.0
django-braces==1.14.0
django-cleanup==5.1.0
django-cors-headers==3.4.0
django-crum==0.7.7
django-extensions==3.1.3
django-import-export==2.4.0
django-oauth-toolkit==1.4.0
django-phonenumber-field==5.0.0
django-polymorphic==3.0.0
django-redis==4.12.1
django-rest-framework==0.1.0
django-rest-passwordreset==1.1.0
django-rest-polymorphic==0.1.9
django-storages==1.11.1
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.4.0
drf-nested-forms==0.2.1
et-xmlfile==1.0.1
fcm-django==0.3.10
google-api-core==1.25.1
google-auth==1.24.0
google-cloud-core==1.5.0
google-cloud-storage==1.35.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
gunicorn==19.10.0
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
numpy==1.20.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
openpyxl-image-loader==1.0.5
pandas==1.2.3
phonenumbers==8.12.21
Pillow==8.0.1
protobuf==3.14.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-decouple==3.4
python-memcached==1.59
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
social-auth-app-django==4.0.0
sqlparse==0.3.1
tablib==2.0.0
toml==0.10.1
urllib3==1.26.3
xlrd==1.2.0
xlwt==1.3.0